Convert Feature interface to abstract class
