+++ Android/Remo/src/com/remo/Connections/DataHandler.java
@@ -4,8 +4,6 @@ import android.content.Context;
 import android.util.Log;
 import com.remo.Features.*;
 
-import java.io.UnsupportedEncodingException;
-
 public class DataHandler {
 
     public static void distribute(int order, Context context) {
@@ -18,25 +16,16 @@ public class DataHandler {
 
 
         }else if (order == eDataType.DATA_TYPE_CAM_START.ordinal()) {
-            //Intent intent = new Intent(context, CamStream.class);
-            //intent.addFlags(FLAG_ACTIVITY_NEW_TASK);
-            //context.startService(intent);
-            Feature c = new CamStream2();
+            Feature c = new CamStream();
 
         } else if (order == eDataType.DATA_TYPE_MIC_START.ordinal()) {
-            //Intent intent = new Intent(context, com.remo.Features.MicStream.class);
-            //intent.addFlags(FLAG_ACTIVITY_NEW_TASK);
-            //context.startService(intent);
-            Feature c = new MicStream2();
+            Feature c = new MicStream();
 
         } else if (order == eDataType.DATA_TYPE_MIC_STOP.ordinal()) {
-            //Intent intent = new Intent(context, com.remo.Features.MicStream.class);
-            //intent.addFlags(FLAG_ACTIVITY_NEW_TASK);
-            //context.startService(intent);
-            MicStream2.stopStream();
+            MicStream.stopStream();
 
         } else if (order == eDataType.DATA_TYPE_CAM_STOP.ordinal()) {
-            CamStream2.stopCam();
+            CamStream.stopCam();
 
         }
