+++ Android/Remo/src/com/remo/Features/Feature.java
@@ -6,19 +6,19 @@ import com.remo.Connections.TCP_Transceiver;
  * Created by Mohamed on 2/1/2018.
  */
 
-public interface Feature {
+public abstract class Feature {
     boolean isRootRequired = false;
     int minSDK = 1;
     String Name = "";
     //TCP_Transceiver tcp = new TCP_Transceiver(false);
     boolean isMaainConn = false;
 
-    void connect();
+    abstract void connect();
 
-    void sendPacket(byte[] data);
+    abstract void sendPacket(byte[] data);
 
-    void reportError(String error);
+    abstract void reportError(String error);
 
-    void disconnect();
+    abstract void disconnect();
 
 }
